@@ -1,8 +1,6 @@
 module github.com/snple/beacon
 
-go 1.23.0
-
-toolchain go1.24.0
+go 1.24.0
 
 require (
 	github.com/BurntSushi/toml v1.4.0
